+++ services/sidecar/Makefile
@@ -2,11 +2,6 @@ GO_CMD=go
 GO_BUILD=$(GO_CMD) build
 CMD_PATH=./cmd/sidecar/main.go
 
-.PHONY: wire-generate # Generate Wire bindings
-wire-generate:
-	cd internal/service/di ;\
-	wire
-
 .PHONY: help # Help - list of targets with descriptions
 help:
 	@echo ''
@@ -15,3 +10,8 @@ help:
 	@echo ''
 	@echo 'Targets:'
 	@grep '^.PHONY: .* #' Makefile | sed 's/\.PHONY: \(.*\) # \(.*\)/ \1\t\2/' | expand -t20
+
+.PHONY: wire-generate # Generate Wire bindings
+wire-generate:
+	cd internal/service/di ;\
+	wire
