+++ services/sidecar/internal/service/di/cli_provider.go
@@ -1,0 +1,10 @@
+package di
+
+import (
+	"github.com/google/wire"
+	"github.com/specularL2/specular/services/sidecar/internal/sidecar/api"
+)
+
+var CliProvider = wire.NewSet( //nolint:gochecknoglobals
+	api.NewCli,
+)
