+++ services/sidecar/internal/service/di/inject.go
@@ -13,6 +13,7 @@ func SetupApplication() (*Application, func(), error) {
 	panic(wire.Build(wire.NewSet(
 		CommonProvider,
 		ConfigProvider,
+		CliProvider,
 		wire.Struct(new(Application), "*"))),
 	)
 }
@@ -20,6 +21,7 @@ func SetupApplication() (*Application, func(), error) {
 func SetupApplicationForIntegrationTests(cfg *config.Config) (*TestApplication, func(), error) {
 	panic(wire.Build(wire.NewSet(
 		CommonProvider,
+		CliProvider,
 		wire.Struct(new(Application), "*"),
 		wire.Struct(new(TestApplication), "*"))),
 	)
